Reorganise class architecture in DynaWaltz
