+++ dynawaltz-dsl/src/test/java/com/powsybl/dynawaltz/dsl/DynaWaltzGroovyEventModelsSupplierTest.java
@@ -14,8 +14,8 @@ import com.powsybl.dynamicsimulation.groovy.EventModelGroovyExtension;
 import com.powsybl.dynamicsimulation.groovy.GroovyEventModelsSupplier;
 import com.powsybl.dynamicsimulation.groovy.GroovyExtension;
 import com.powsybl.dynawaltz.DynaWaltzProvider;
-import com.powsybl.dynawaltz.events.AbstractBlackBoxEventModel;
-import com.powsybl.dynawaltz.events.EventQuadripoleDisconnection;
+import com.powsybl.dynawaltz.dynamicmodels.nonstaticref.events.AbstractBlackBoxEventModel;
+import com.powsybl.dynawaltz.dynamicmodels.nonstaticref.events.EventQuadripoleDisconnection;
 import com.powsybl.dynawaltz.dsl.events.EventQuadripoleDisconnectionGroovyExtension;
 import com.powsybl.dynawaltz.dsl.events.EventSetPointBooleanGroovyExtension;
 import com.powsybl.iidm.network.Identifiable;
