+++ dynawaltz-dsl/src/test/resources/ieee14-currentlimitautomaton/powsybl-inputs/dynamicModels.groovy
@@ -5,10 +5,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-
-import com.powsybl.dynawaltz.automatons.CurrentLimitAutomaton
 import com.powsybl.iidm.network.Branch
-import com.powsybl.iidm.network.Line
 import com.powsybl.iidm.network.Load
 import com.powsybl.iidm.network.Generator
