+++ dynawaltz/src/main/java/com/powsybl/dynawaltz/dynamicmodels/nonstaticref/automatons/CurrentLimitAutomaton.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.dynawaltz.automatons;
+package com.powsybl.dynawaltz.dynamicmodels.nonstaticref.automatons;
 
 import javax.xml.stream.XMLStreamException;
 import javax.xml.stream.XMLStreamWriter;
@@ -12,6 +12,7 @@ import javax.xml.stream.XMLStreamWriter;
 import com.powsybl.commons.PowsyblException;
 import com.powsybl.dynawaltz.DynaWaltzContext;
 import com.powsybl.dynawaltz.dynamicmodels.*;
+import com.powsybl.dynawaltz.dynamicmodels.nonstaticref.network.LineModel;
 import com.powsybl.iidm.network.*;
 import org.apache.commons.lang3.tuple.Pair;
