+++ dynawaltz/src/main/java/com/powsybl/dynawaltz/dynamicmodels/nonstaticref/events/BlackBoxEventModel.java
@@ -4,17 +4,15 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.dynawaltz.dynamicmodels;
+package com.powsybl.dynawaltz.dynamicmodels.nonstaticref.events;
+
+import com.powsybl.dynawaltz.dynamicmodels.BlackBoxModel;
 
 /**
  * @author Florian Dupuy <florian.dupuy at rte-france.com>
  */
-public interface BusModel extends BlackBoxModel {
-    String getId();
+public interface BlackBoxEventModel extends BlackBoxModel {
 
-    String getTerminalVarName();
+    String getEventModelId();
 
-    String getSwitchOffSignalVarName();
-
-    String getNumCCVarName();
 }
