+++ dynawaltz/src/main/java/com/powsybl/dynawaltz/dynamicmodels/nonstaticref/events/EventQuadripoleDisconnection.java
@@ -4,12 +4,12 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.dynawaltz.events;
+package com.powsybl.dynawaltz.dynamicmodels.nonstaticref.events;
 
 import com.powsybl.commons.PowsyblException;
 import com.powsybl.dynawaltz.DynaWaltzContext;
 import com.powsybl.dynawaltz.dynamicmodels.BlackBoxModel;
-import com.powsybl.dynawaltz.dynamicmodels.LineModel;
+import com.powsybl.dynawaltz.dynamicmodels.nonstaticref.network.LineModel;
 import com.powsybl.iidm.network.Branch;
 import org.apache.commons.lang3.tuple.Pair;
