+++ dynawaltz/src/main/java/com/powsybl/dynawaltz/dynamicmodels/nonstaticref/events/EventSetPointBoolean.java
@@ -4,12 +4,12 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.dynawaltz.events;
+package com.powsybl.dynawaltz.dynamicmodels.nonstaticref.events;
 
 import com.powsybl.commons.PowsyblException;
 import com.powsybl.dynawaltz.DynaWaltzContext;
 import com.powsybl.dynawaltz.dynamicmodels.BlackBoxModel;
-import com.powsybl.dynawaltz.dynamicmodels.GeneratorModel;
+import com.powsybl.dynawaltz.dynamicmodels.staticref.generators.GeneratorModel;
 import org.apache.commons.lang3.tuple.Pair;
 
 import java.util.List;
