+++ dynawaltz/src/main/java/com/powsybl/dynawaltz/dynamicmodels/staticref/generators/GeneratorSynchronousModel.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.dynawaltz.dynamicmodels;
+package com.powsybl.dynawaltz.dynamicmodels.staticref.generators;
 
 /**
  * @author Florian Dupuy <florian.dupuy at rte-france.com>
