+++ dynawaltz/src/main/java/com/powsybl/dynawaltz/dynamicmodels/staticref/loads/LoadAlphaBeta.java
@@ -4,9 +4,11 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.dynawaltz.dynamicmodels;
+package com.powsybl.dynawaltz.dynamicmodels.staticref.loads;
 
 import com.powsybl.commons.PowsyblException;
+import com.powsybl.dynawaltz.dynamicmodels.BlackBoxModel;
+import com.powsybl.dynawaltz.dynamicmodels.nonstaticref.network.BusModel;
 import org.apache.commons.lang3.tuple.Pair;
 
 import java.util.Arrays;
