+++ dynawaltz/src/main/java/com/powsybl/dynawaltz/xml/DydXml.java
@@ -11,7 +11,7 @@ import com.powsybl.commons.exceptions.UncheckedXmlStreamException;
 import com.powsybl.dynawaltz.DynaWaltzContext;
 import com.powsybl.dynawaltz.dynamicmodels.BlackBoxModel;
 import com.powsybl.dynawaltz.dynamicmodels.MacroConnector;
-import com.powsybl.dynawaltz.events.BlackBoxEventModel;
+import com.powsybl.dynawaltz.dynamicmodels.nonstaticref.events.BlackBoxEventModel;
 
 import javax.xml.stream.XMLStreamException;
 import javax.xml.stream.XMLStreamWriter;
