+++ dynawaltz/src/test/java/com/powsybl/dynawaltz/xml/DynamicModelsXmlTest.java
@@ -9,7 +9,7 @@ package com.powsybl.dynawaltz.xml;
 import com.powsybl.dynamicsimulation.DynamicSimulationParameters;
 import com.powsybl.dynawaltz.DynaWaltzContext;
 import com.powsybl.dynawaltz.DynaWaltzParameters;
-import com.powsybl.dynawaltz.dynamicmodels.GeneratorFictitious;
+import com.powsybl.dynawaltz.dynamicmodels.staticref.generators.GeneratorFictitious;
 import org.junit.Test;
 import org.xml.sax.SAXException;
